Fail CI pipeline when clang-tidy fails (#461)
- Exit with correct code when linting fails
- Fix linter errors in the code

+++ editor/lib/quoll/editor/scene/renderer/MousePickingGraph.cpp
@@ -329,7 +329,7 @@ void MousePickingGraph::createRenderGraph() {
         for (const auto &geometry : meshData.drawData->geometries) {
           commandList.drawIndexed(geometry.numIndices, indexOffset,
                                   vertexOffset, numInstances, instanceStart);
-          vertexOffset += geometry.numVertices;
+          vertexOffset += static_cast<i32>(geometry.numVertices);
           indexOffset += geometry.numIndices;
         }
         instanceStart += numInstances;
@@ -360,7 +360,7 @@ void MousePickingGraph::createRenderGraph() {
         for (const auto &geometry : meshData.drawData->geometries) {
           commandList.drawIndexed(geometry.numIndices, indexOffset,
                                   vertexOffset, numInstances, instanceStart);
-          vertexOffset += geometry.numVertices;
+          vertexOffset += static_cast<i32>(geometry.numVertices);
           indexOffset += geometry.numIndices;
         }
         instanceStart += numInstances;
